# schemas/openapi/markets-v1.yaml
openapi: 3.1.0
info:
  title: RetroPick Markets API
  version: 1.3.0
  description: |
    Public, Polymarket-native BFF shared by future web and Android clients.
    Polymarket remains the venue and settlement authority.

    Money representations (no binary floating point):
    - DecimalString — unsigned decimal strings for prices, probabilities, and book sizes (PHASE-1 read).
    - MoneyAmount — integer base units plus currency and decimals for balances, funding, and fees (PHASE-2+).

    PHASE-3 order submit, cancel, and list paths are spec-frozen in this document.
    Runtime handlers remain gated by GET /markets/capabilities features.order_submit until MKT-P3-002 wiring.

    Every upstream-derived response carries provenance and freshness metadata.
servers:
  - url: /api/v1
...
paths:
...
  /markets/me/wallets:
    get:
      tags: [Markets]
      operationId: listMyWallets
      summary: List linked Polymarket account wallets for the authenticated session
      description: |
        Returns the session signer EOA and any linked Polymarket trading wallets
        (proxy, Safe, or Deposit Wallet). signerAddress and accountWallet MUST remain
        distinct fields per ADR-003. Empty wallets means not yet linked — not an error.
      x-phase: 2
      security:
        - MarketsSession: []
      responses:
        "200":
          description: Signer and linked account wallets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletsListResponse"
              examples:
                linkedProxy:
                  summary: Proxy linked — signer differs from account wallet
                  value:
                    schemaVersion: "1"
                    signerAddress: "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                    wallets:
                      - accountWallet: "0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                        walletType: GNOSIS_SAFE
                        linkStatus: linked
                        isPrimary: true
                        chainId: 137
                    checkedAt: "2026-08-09T10:00:00Z"
                notLinked:
                  summary: No linked account wallet yet
                  value:
                    schemaVersion: "1"
                    signerAddress: "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                    wallets: []
                    checkedAt: "2026-08-09T10:00:00Z"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /markets/me/wallets/link:
    post:
      tags: [Markets]
      operationId: linkExistingWallet
      summary: Connect an existing Polymarket account wallet to the session signer
      description: |
        Persists a connect-existing linkage between the session signer EOA and a
        client-supplied Polymarket trading wallet. The BFF never generates or invents
        accountWallet addresses. Response is a bare LinkedWallet object. Responses
        MUST NOT be cached (private, no-store).
      x-phase: 2
      security:
        - MarketsSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkExistingWalletRequest"
            examples:
              linkProxy:
                summary: Link existing Gnosis Safe proxy wallet
                value:
                  accountWallet: "0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                  walletType: GNOSIS_SAFE
                  isPrimary: true
      responses:
        "200":
          description: Linked account wallet persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkedWallet"
              examples:
                linkedProxy:
                  summary: Proxy linked — signer differs from account wallet
                  value:
                    accountWallet: "0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                    walletType: GNOSIS_SAFE
                    linkStatus: linked
                    isPrimary: true
                    chainId: 137
        "400":
          description: Invalid wallet address or malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                invalidRequest:
                  value:
                    error:
                      code: invalid_request
                      message: invalid wallet address
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Wallet linkage conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                conflict:
                  value:
                    error:
                      code: conflict
                      message: wallet linkage conflict
        "503":
          description: Wallet linking is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                serviceUnavailable:
                  value:
                    error:
                      code: service_unavailable
                      message: wallet linking is not configured
  /markets/account-wallet/preview:
    post:
      tags: [Markets]
      operationId: previewAccountWallet
      summary: Preview metadata for deposit-wallet deploy or link-existing
      description: |
        Returns preview metadata for the requested account-wallet action. Does not
        call upstream relayer HTTP, does not persist linkage, and exposes no relayer
        secrets. When action is omitted, defaults to deploy_deposit_wallet. Responses
        MUST NOT be cached (private, no-store).
      x-phase: 2
      security:
        - MarketsSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountWalletPreviewRequest"
            examples:
              deployDepositWallet:
                summary: Preview deposit wallet deploy
                value:
                  action: deploy_deposit_wallet
              emptyBody:
                summary: Default action when body is empty object
                value: {}
      responses:
        "200":
          description: Preview metadata for client confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountWalletPreviewResponse"
              examples:
                deployDepositWallet:
                  summary: Deploy Deposit Wallet preview
                  value:
                    schemaVersion: "1"
                    signerAddress: "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                    action: deploy_deposit_wallet
                    chainId: 137
                    message: Sign and deploy Deposit Wallet via upstream relayer; BFF persists the deployed address on relay.
        "400":
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Wallet linking is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /markets/account-wallet/relay:
    post:
      tags: [Markets]
      operationId: relayAccountWallet
      summary: Persist client-supplied deployed Deposit Wallet address
      description: |
        Persists a deployed Deposit Wallet address supplied by the client after
        upstream relayer deploy. The BFF never generates or invents accountWallet
        addresses. Idempotent upsert at the store layer. Responses MUST NOT be cached
        (private, no-store).
      x-phase: 2
      security:
        - MarketsSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountWalletRelayRequest"
            examples:
              relayDepositWallet:
                summary: Persist deployed Deposit Wallet
                value:
                  accountWallet: "0xdddddddddddddddddddddddddddddddddddddddd"
                  isPrimary: true
      responses:
        "200":
          description: Deposit Wallet linkage persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountWalletRelayResponse"
              examples:
                relayed:
                  summary: Relay persisted deposit wallet
                  value:
                    schemaVersion: "1"
                    signerAddress: "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                    wallet:
                      accountWallet: "0xdddddddddddddddddddddddddddddddddddddddd"
                      walletType: DEPOSIT_WALLET
                      linkStatus: linked
                      isPrimary: true
                      chainId: 137
        "400":
          description: Missing or invalid account wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Wallet linkage conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Wallet linking is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    MarketsSession:
      type: apiKey
      in: cookie
      name: retropick_markets_session
      description: HttpOnly session cookie issued by Markets auth (MKT-P2-005).
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID for mutating POST idempotency (24h replay window).
      schema:
        type: string
        format: uuid
...
  schemas:
    SchemaVersion:
      type: string
      const: "1"
...
    WalletType:
      type: string
      enum: [EOA, POLY_PROXY, GNOSIS_SAFE, DEPOSIT_WALLET]
      description: Polymarket wallet type per EV-009.
    LinkStatus:
      type: string
      enum: [linked, pending_verification]
    LinkedWallet:
      type: object
      additionalProperties: false
      required: [accountWallet, walletType, linkStatus, isPrimary, chainId]
      properties:
        accountWallet:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: Polymarket trading wallet holding collateral and positions.
        walletType:
          $ref: "#/components/schemas/WalletType"
        linkStatus:
          $ref: "#/components/schemas/LinkStatus"
        isPrimary:
          type: boolean
        chainId:
          type: integer
          minimum: 1
          examples: [137]
    AccountWalletAction:
      type: string
      enum: [link_existing, deploy_deposit_wallet]
      description: Intended account-wallet operation for preview.
    LinkExistingWalletRequest:
      type: object
      additionalProperties: false
      required: [accountWallet, walletType]
      properties:
        accountWallet:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: Existing Polymarket trading wallet to link.
        walletType:
          $ref: "#/components/schemas/WalletType"
        linkStatus:
          $ref: "#/components/schemas/LinkStatus"
        isPrimary:
          type: boolean
          default: true
        chainId:
          type: integer
          minimum: 1
          examples: [137]
        linkageProofHash:
          type: string
          description: Optional signature-challenge hash; no relayer API keys.
    AccountWalletPreviewRequest:
      type: object
      additionalProperties: false
      properties:
        action:
          $ref: "#/components/schemas/AccountWalletAction"
          default: deploy_deposit_wallet
    AccountWalletPreviewResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, signerAddress, action, chainId, message]
      properties:
        schemaVersion:
          $ref: "#/components/schemas/SchemaVersion"
        signerAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: Session signer EOA (ADR-003); never from request body.
        action:
          $ref: "#/components/schemas/AccountWalletAction"
        chainId:
          type: integer
          minimum: 1
          examples: [137]
        message:
          type: string
          description: Human-readable preview summary for client confirmation modal.
    AccountWalletRelayRequest:
      type: object
      additionalProperties: false
      required: [accountWallet]
      properties:
        accountWallet:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: Deployed Deposit Wallet address supplied by the client.
        chainId:
          type: integer
          minimum: 1
          examples: [137]
        isPrimary:
          type: boolean
          default: true
        linkageProofHash:
          type: string
          description: Optional signature-challenge hash; no relayer API keys.
    AccountWalletRelayResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, signerAddress, wallet]
      properties:
        schemaVersion:
          $ref: "#/components/schemas/SchemaVersion"
        signerAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: Session signer EOA (ADR-003); never from request body.
        wallet:
          $ref: "#/components/schemas/LinkedWallet"
    WalletsListResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, signerAddress, wallets, checkedAt]
      properties:
        schemaVersion:
          $ref: "#/components/schemas/SchemaVersion"
        signerAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          description: EOA that signs L1 auth and EIP-712 payloads (ADR-003).
        wallets:
          type: array
          items:
            $ref: "#/components/schemas/LinkedWallet"
        checkedAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
      examples:
        - error:
            code: not_found
            message: Resource not found
            requestId: req-conformance-1
...